# api/docs/swagger.yaml
definitions:
...
  tx.EvmTxCountResponse:
    properties:
      count:
        type: integer
        x-order:0: true
    type: object
info:
  contact: {}
paths:
...
  /indexer/nft/v1/txs/{collection_addr}/{token_id}:
    get:
      consumes:
      - application/json
      description: Get NFT transactions for a specific collection and token ID
      parameters:
      - description: Collection address
        in: path
        name: collection_addr
        required: true
        type: string
      - description: Token ID
        in: path
        name: token_id
        required: true
        type: string
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses: {}
      summary: Get NFT transactions by collection and token ID
      tags:
      - NFT
  /indexer/tx/v1/evm-txs:
    get:
      consumes:
      - application/json
      description: Get a list of EVM transactions with pagination
      parameters:
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get EVM transactions
      tags:
      - Evm Transactions
  /indexer/tx/v1/evm-txs/{tx_hash}:
    get:
      consumes:
      - application/json
      description: Get a specific EVM transaction by its hash
      parameters:
      - description: Transaction hash
        in: path
        name: tx_hash
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get EVM transaction by hash
      tags:
      - Evm Transactions
  /indexer/tx/v1/evm-txs/by_account/{account}:
    get:
      consumes:
      - application/json
      description: Get EVM transactions associated with a specific account
      parameters:
      - description: Account address
        in: path
        name: account
        required: true
        type: string
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get EVM transactions by account
      tags:
      - Evm Transactions
  /indexer/tx/v1/evm-txs/by_height/{height}:
    get:
      consumes:
      - application/json
      description: Get EVM transactions at a specific block height
      parameters:
      - description: Block height
        in: path
        name: height
        required: true
        type: integer
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get EVM transactions by height
      tags:
      - Evm Transactions
  /indexer/tx/v1/evm-txs/count:
    get:
      consumes:
      - application/json
      description: Get the total number of EVM transactions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tx.EvmTxCountResponse'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get EVM transaction count
      tags:
      - Evm Transactions
...
swagger: "2.0"